fix warnings

+++ c_src/modules/evision_backend/broadcast.h
@@ -77,7 +77,7 @@ static ERL_NIF_TERM evision_cv_mat_broadcast_to(ErlNifEnv *env, int argc, const 
             int64_t ndims = to_shape.size();
             std::vector<int> src_shape(ndims);
 
-            int diff_dims = ndims;
+            int diff_dims = (int)ndims;
             if (force_src_shape.size() > 0) {
                 diff_dims -= force_src_shape.size();
             } else {
@@ -125,7 +125,7 @@ static ERL_NIF_TERM evision_cv_mat_broadcast_to(ErlNifEnv *env, int argc, const 
             broadcast(img, dst_data, tmp_data, src_shape, to_shape, elem_size);
 
             int type = img.type() & CV_MAT_DEPTH_MASK;
-            Mat result = Mat(ndims, to_shape.data(), type, dst_data);
+            Mat result = Mat((int)ndims, to_shape.data(), type, dst_data);
             result = result.clone();
             enif_free((void *)dst_data);
             enif_free((void *)tmp_data);
